Add configuration documentation, no longer use a dictionary for Browser's configuration.
In the process of cleaning up the configuration module, moved away
from using a dictionary for internal storage of Browser's configuration.
Using a dictionary meant that keys could be specified in the
configuration that didn't do anything, and didn't raise an error.
The downside of this approach is pylint complaining about star-args.